Corrigindo testes e exportanto a função para a pasta utils
